Compute covariance matrix of incomplete data using multiple imputation

Compute covariance matrix of incomplete data using multiple imputation. For multiple imputation, Multivariate Imputation by Chained Equations (MICE) from the mice package is used. Also provides variance explained by different numbers of factors with Fieller (parametric) or bootstrap (nonparametric) confidence intervals.

mifa(data, cov_var, n_factors, ci = FALSE, conf = 0.95, n_boot = 1000, ...)

Arguments

data
A data frame with missing values coded as NA.

cov_var
Variables in data to calculate the covariance matrix for. Supports (tidy selection)dplyr::select(). This allows to select variables that are used for the imputations of missing values, but not the calculations of the covariance matrix. This is especially useful when there are categorical predictors that can improve the imputation of the response variables, but for which covariance cannot be calculated. By default, all variables in data are used for both, the imputation and the covariance matrix. Note: Variables and rows used for the imputation can be configured using the .... See also mice::mice().

n_factors
Integer or integer vector indicating number of factors for which proportion of explained variance (and optionally confidence intervals) should be indicated. Defaults to all possible numbers of factors.

ci
A character string or character vector indicating which types of confidence intervals should be constructed. If "bootstrap", "fieller", or "both", the corresponding intervals are computed. If FALSE (the default) no confidence intervals will be computed.

conf
Confidence level for constructing confidence intervals. The default is .95, i.e., a 95% confidence interval.

n_boot
number of bootstrap samples to use for the bootstrap confidence intervals. The default is 1000.

...
Named arguments that will be passed to mice::mice(), such as

m Number of generated imputations. The default is 5.

maxit Number of iterations for each imputation. The default is 5.

method Method used for imputation.

printFlag Whether to print diagnostic information to console. The default is TRUE). See mice::mice() for all options.

Value

A list:

cov_combined

The estimated covariance matrix of the incomplete data, based on the combined covariance matrices of imputed data sets.

cov_imputations

A list containing the estimated covariance matrixes for all imputed data sets.

var_explained

A data frame containing the estimated proportions of explained variance for each of specified n.factor components. Depending o n ci, it will also contain the estimated Fieller's (parametric) and/or bootstrap (nonparametric) confidence interval for the proportion of variance explained by the different numbers of factors defined by n_factors.

References

Nassiri, V., Lovik, A., Molenberghs, G., & Verbeke, G. (2018). On using multiple imputation for exploratory factor analysis of incomplete data. Behavioral Research Methods 50, 501–517. https://doi.org/10.3758/s13428-017-1013-4

See also