Fieller confidence intervals for explained variance

Computes parametric confidence intervals for proportion of explained variance for given numbers of principal components using Fieller's method. Note that by setting ci = TRUE in mifa(), this confidence interval can be computed as well.

mifa_ci_fieller(cov_imps, n_pc, conf = 0.95, N)

Arguments

cov_imps

List containing the estimated covariance matrix within each imputed data. One can use cov_imputations returned by mifa().

n_pc

Integer or integer vector indicating number of principal components (eigenvectors) for which explained variance (eigenvalues) should be obtained and for which confidence intervals should be computed. Defaults to all principal components, i.e., the number of variables in the data.

conf

Confidence level for constructing confidence intervals. The default is .95 that is, 95% confidence intervals.

N

A scalar specifying sample size.

Value

A data frame containing confidence intervals for n_pc principal components.

Details

Normally, this function does not need to be called directly. Instead, use mifa(..., ci = "fieller").

References

Fieller, E. C. (1954). Some problems in interval estimation. Journal of the Royal Statistical Society. Series B (Methodological): 175-185.

See also

Examples

# \donttest{
if(requireNamespace("psych")) {
  data <- psych::bfi[, 1:25]
  mi <- mifa(data, print = FALSE)
  mifa_ci_fieller(mi$cov_imputations, n_pc = 3:8, N = nrow(data))
}
#>   n_pc     lower     upper
#> 1    3 0.4025181 0.4195041
#> 2    4 0.4732084 0.4889571
#> 3    5 0.5340087 0.5484998
#> 4    6 0.5780892 0.5915733
#> 5    7 0.6148407 0.6274452
#> 6    8 0.6504145 0.6621411
# }